// front/src/app/api-client/model/routine-run-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * État d\'une routine pour un jour (les coches repartent de zéro chaque jour)
 */
export interface RoutineRunDTO { 
    date?: string;
    /**
     * Ids des étapes cochées
     */
    etapesCochees?: Array<number>;
    /**
     * Toutes les étapes sont cochées
     */
    terminee?: boolean;
    /**
     * Moment où la dernière étape a été cochée
     */
    termineeLe?: string | null;
    /**
     * La récompense du jour (mini-jeu) a été jouée
     */
    recompenseJouee?: boolean;
}